Gets rid of the newline / nobreak feature.
With the newline / nobreak feature a user could align multiple questions horizontally.

Apart from the fact that I'm not sure if it is even used I would rather implement it by using css instead of <br/> calls in different widgets. As it was it didn't really work well with twitter bootstrap, so I'd rather take it out of the equation for now and concentrate on good default rendering.