Remove dead link, correct and comment other link
The one alternative cannot be found on the link, the trac system is down, and the website of the main domain is redirected to another company, which doesn't mention the questionnaire. Also correct the link to the other alternative which was not correct in markup, and add comment that the software hasn't had a change since 2009.